requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.0
vertexai==1.38.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.0
python-dotenv==1.0.0
db-dtypes==1.1.1